server/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.32.3
PyJWT==2.8.0
beautifulsoup4==4.12.3
bleach==5.0.1
fastapi==0.104.1
